fastrpc-sepolicy:Enable unlink of tmpfs_t
fastrpc uses LRM libraries to communicate with each other. lrmc creates
queue files for this purpose. In normal cases, fastrpc_lib has
destructor which calls lrmc_disconnect to cleanup these queue files. But
in abnormal exit, the fallback implemented is for lrms lib which is
linked to fastrpc-rm server to cleanup these files. But, fastrpc-rm is
unable to cleanup due to sepolicy enforcement. The lrmc queue files are
created under tmpfs_t context, so we are enabling fastrpc_rm_t
permission to unlink tmpfs_t. We are aware of the security concerns with
this change as tmpfs_t is a generic context, this is just a workaround
until proper fix is in place.

Change-Id: I1fd80fc72a711a3a245b8a85c9306b543f921d76
Signed-off-by: Vinayaka B M <vinaybm@qti.qualcomm.com>
(cherry picked from commit 24c37ece1849fef32cfd44a6bab416d28c47dd21)